BMI Calculator
I created a BMI calculator through python.  BMI stands for Body Mass Index, which is a measure of body fat that is based on the person's height and weight. This can be applied to adults of any gender, men or women.  My code first starts with an input function asking for the person's name then I create two more input functions but with a float. This asking for the person's weight and height. Then I created a variable with the BMI formula. Then I used the if function staying if the BMI is less than 25 the person is underweight. That's because on the BMI chart 25 is normal. Try my code out!